Split user.ts into auth.ts and me.ts
